LayoutTests/http/tests/security/contentSecurityPolicy/worker-blob-inherits-csp-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/post-message.js because it does not appear in the script-src directive of the Content Security Policy.

PASS Blob worker inherits CSP 
